Added FLAC and OggVorbis parsing / functionality.
Lots of new changes. Functions now take a tuple for each format choice
which includes the list of files in that format (mp3, ogg, flac), the
associated mutagen function, and the ending to use.

Directory structure is set to append the format to the album name in
case you have the same album 3 or 4 ways. Also makes sure you don't have
different formats in one folder.